+++ pinot-connectors/pinot-flink-connector/README.md
@@ -42,6 +42,6 @@ execEnv.execute();
 For more examples, please see `src/main/java/org/apache/pinot/connector/flink/FlinkQuickStart.java`
 
 ## Notes for backfilling upsert table
- - To correctly partition the output segments by the primary key, the Flink job *must* also include the partitionBeyKey operator before the Sink operator
+ - To correctly partition the output segments by the primary key, the Flink job *must* also include the partitionByKey operator before the Sink operator
  - The parallelism of the job *must* be set the same as the number of partitions of the Pinot table, so that the sink in each task executor can generate the segment of same partitions.
  - It’s important to plan the resource usage to avoid capacity issues such as out of memory. In particular, Pinot sink has an in-memory buffer of records, and it flushes when the threshold is reached. Currently, the threshold on the number of records is supported via the config of `segmentFlushMaxNumRecords`. In the future, we could add other types of threshold such as the memory usage of the buffer.
